+++ public/front-js/main.js
@@ -19,33 +19,5 @@ window.onload = function(){
     handsPic.classList.add('goRight')
 }
 
-// 슬라이드 구현
-const preBtn = document.querySelector('button#prebtn')
-const nextBtn = document.querySelector('#nextbtn')
-const slider = document.querySelector('.leftSlider')
-
-function slideshow(){
-    slider.style.transform = `translateX(${(click-1)*30}%)`
-}
-
-let click = 0;
-nextBtn.addEventListener('click',()=>{
-    click++;
-    if(click=2){
-        click=0
-    }
-    slider.style.transform = `translateX(${(click-1)*30}%)`
-})
-
-// let click = 1;
-// nextBtn.onclick = function(){
-//     click--;
-//     if(click <)
-
-
-
-
-
-
 
 })
