+++ routes/index.js
@@ -2,11 +2,9 @@ const express = require(`express`)
 const app = express()
 const router = express.Router()
 
-
 const boardRouter = require(`./board`)
 const userRouter = require('./user/index.js')
 const mainController = require(`./main.Controller`)
-
 
 app.use(express.urlencoded({extended:true}))
