+++ routes/main.Controller.js
@@ -1,3 +1,7 @@
+
+const express = require('express')
+const router = express.Router()
+
 exports.main =
     (req,res)=>{
         res.render(`main`)
