+++ routes/user/loginController.js
@@ -1,25 +1,35 @@
 const res = require("express/lib/response")
+const router = require(".")
 const { connect } = require(".")
 const pool = require("../../db")
 const { alertmove } = require("../util/alertmove")
 
-let loginUser = function(id, pw) {
-    let isLogin = false;
-    let loginSql = `select * from userAccount where id = '${id}' and pw = '${pw}' ;`
-    pool.getConnection((err, conn) => {
-        conn.query(loginSql,(err, result) => {
-            if(!err) {
-                if(result[0].id != undefined) { 
-                    console.log('비번오케이')
-                    return isLogin = true;
+// 로그인 get 
+exports.login = (req,res) => {
+    res.render('user/login.html')
+} 
+
+// 로그인 post
+exports.loginAction = (req,res) => {
+    const {id,pw} = req.body
+    const loginData = {id,pw}
+    let loginSql = `select * from userAccount where id = '${id}' and pw = '${pw}';`
+    pool.getConnection((err,conn)=>{
+        conn.query(loginSql, (err,result)=>{
+            if(!err){
+                if(result.length != 0){
+                    req.session.user = {...loginData}
+                    const {user} = req.session
+                    console.log(req.session)
+                    res.render('main.html',{
+                        user,
+                    })
+                }
+                else{
+                    res.send(alertmove('/user/login','비번확인바람'))
                 }
             }
-            else {throw err}
         })
         conn.release();
     })
-    return true;
 }
-
-module.exports = loginUser
-
