+++ server.js
@@ -1,7 +1,6 @@
 const express = require('express')
 const nunjucks = require('nunjucks')
 const router = require(`./routes`)
-
 const app = express()
 const port = process.env.PORT || 3000
 
@@ -10,10 +9,24 @@ nunjucks.configure('views',{express:app,})
 app.use(express.static('public'))
 app.use(express.urlencoded({extended:true,}))
 
+//yj 추가함 (세션)
+const session = require('express-session')
+const Memorystore = require('memorystore')(session)
+//yj 추가함 (세션)
+let sessionObj = {
+    secret: 'admin',
+    resave: false,
+    saveUninitialized:true,
+    store:new Memorystore({checkperiod: 30*60*1000}),
+    cookie: {
+        maxAge: 30*60*1000
+    }
+}
+
+app.use(session(sessionObj))
 app.use(router)
 
 
-//
 
 // 관리자 라우터 분리하기
 app.get('/admin', (req,res)=>{
