+++ views/_navbar.html
@@ -9,7 +9,11 @@
         <li id="user"><a href="/user/login">MEMBERS</a>
             <ul class="snb">
                 <li><a href="/user/join">JoinUs</a></li>
+                {% if user %}
+                <li><a href="/user/logout">Logout</a></li>            
+                {% else %}
                 <li><a href="/user/login">Login</a></li>
+                {% endif %}
                 <li><a href="/user/profile">Profile</a></li>
             </ul>
         </li>
