+++ views/user/join.html
@@ -52,11 +52,11 @@
             </li>
             <li id="phone">
                 <span>전화번호</span>
-                <input type="text" name="phone" id="phone1" maxlength="11">
+                <input type="text" name="phone" maxlength="11" placeholder="'-'는 제외하고 입력해주세요">
             </li>
             <li id="mobile">
                 <span>핸드폰번호</span>
-                <input type="text" name="mobile" id="mobile1" maxlength="11">
+                <input type="text" maxlength="11" name="mobile" placeholder="'-'는 제외하고 입력해주세요">
             </li>
         </ul>
         <input type="submit" value="회원가입" id="submitBtn">
